Manim_microservice/requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
Flask-CORS==5.0.1
requests==2.31.0
boto3==1.34.0
gunicorn==21.2.0
docker==6.1.0